% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wesanderson.R
\docType{package} 
\name{JosephsLabColors} 
\alias{JosephsLabColors} 
\alias{JosephsLabColors-package} 
\title{JosephsLabColors} 
\description{ 
list of palettes for Emily Josephs Lab at Michigan State University 
}
\details{
list of palettes for Emily Josephs Lab at Michigan State University
}
